Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/browsers/history/the-location-interface/allow_prototype_cycle_through_location.sub-expected.txt

CONSOLE MESSAGE: SecurityError: Attempted to use a non-registrable domain.

FAIL same-origin, same-window location cycle cyclic __proto__ value
PASS cross-origin location has null prototype
FAIL same-origin, different-window location cycle cyclic __proto__ value
FAIL cross-origin, but joined via document.domain, location cycle Blocked a frame with origin "http://localhost:8800" from accessing a cross-origin frame. Protocols, domains, and ports must match.
